\name{continuousdata.csv}
\alias{continuousdata}
\title{
continuousdata
}
\description{
Example 2 culture continuous data with item difficulty for cctgui() \cr
This data should load as 40 respondents by 40 items, and as continuous data \cr
It is an example of 2 culture data, with 16 missing data values \cr
}
\note{
csv or text data files need not use header or row names \cr
Though respondents should be by the rows, and items by the columns 
}
