\name{hotcold.csv}
\alias{hotcold}
\title{
hotcold
}
\description{
Example 2 culture binary data with item difficulty for cctgui() \cr
This data should load as 23 respondents by 27 items, and as binary data \cr
It is an example of 2 culture data, with 14 missing data values \cr
}
%\usage{
%}
%\details{
%}
%\value{
%}
\note{
csv or text data files need not use header or row names \cr
Though respondents should be by the rows, and items by the columns 
}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
