\name{print.occult}
\alias{print.occult}
\title{
Customizes function printing
}
\description{
Provides function arguments but hides function body
}
\usage{
\method{print}{occult}(x, ...)
}

%\details{
%}
%\references{
%}
\author{
Royce Anders
}
%\note{
%}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{internal}
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
