\name{raterdata.csv}
\alias{raterdata}
\title{
raterdata
}
\description{
Example 1 culture ordinal data with item difficulty for cctgui() \cr
This data should load as 25 respondents by 40 items, and as ordinal data \cr
It is an example of 1 culture data, with 9 missing data values \cr
}
%\usage{
%}
%\details{
%}
%\value{
%}
\note{
csv or text data files need not use header or row names \cr
Though respondents should be by the rows, and items by the columns 
}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
