\name{testdat.csv}
\alias{testdat}
\title{
testdat
}
\description{
Example test data for cctgui() \cr
This data should load as 20 respondents by 25 items, and as dichotomous data \cr
It is an example of 1 culture data \cr
}
%\usage{
%}
%\details{
%}
%\value{
%}
\note{
csv or text data files need not use header or row names \cr
Though respondents should be by the rows, and items by the columns 
}
%\examples{
%}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
